+++ src/main/java/io/probedock/junitee/dependency/DependencyInjector.java
@@ -1,6 +1,6 @@
 package io.probedock.junitee.dependency;
 
-import io.probedock.junitee.generator.InjectDataGenerator;
+import io.probedock.junitee.annotations.InjectDataGenerator;
 import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.HashSet;